Progress #4: Only write irreversible blocks
As of this commit, we only write irreversible blocks to disk, and when
we reopen the database, we rewind the state until the head block is the
last irreversible block.

The unit tests are doing their jobs: several of them broke due to this
change, because the tests which checked that the head block number
didn't change during a restart of the database were invalidated by this
change: the head block number *should* now change in a restart; it
should revert to the last_irreversible_block_num. The tests are updated
to account for this, and all pass again.