Add a template for plugins
This is just an empty template plugin which will serve as a starting
point for new plugins.

Also add a helper script, make_new_plugin.sh, which copies the template
plugin to create your new plugin and automatically renames all of the
files/directories/classes/binaries/whatever. Invoke like so:
./make_new_plugin.sh new_plugin
Afterwards, the new_plugin folder will be ready to go.