Update README.md
Remove discussion of outdated dependencies and update instructions to describe a single-node testnet.

Since the P2P code was the reason we had the outdated dependencies that made setting up a development environment such a pain, we removed it until we can implement a better solution. In the meantime, only single-node testnets are possible.