Ref #15, #19: Progress Un/Locking EOS
Initial work on locking EOS, beginning unlock, and claiming unlocked
EOS.

TODO: Actually split the system contract, and define names for the
different parts so we can check that the required notifications are made.
Also, figure out how to write notify handlers for the native contracts,
to handle these critical balance updates in the staking/unstaking of
EOS. See the #warning TODO comments in the commit for context.