Ref #19: Implement producer scheduling
Calculating the new producer schedule based on votes is now implemented;
however, the voting itself is not. :D But the code builds, and the tests
pass. It's good progress, it's been a good day.

Note that "the tests pass" means little at this stage, since I haven't
written any new tests to exercise the new code. In other words, NEEDS
TESTING!